+++ apps/editor-dev/pages/messages/index.js
@@ -1,0 +1,1 @@
+export { ko_KR, en_US } from "./locales";