+++ apps/editor-dev/pages/messages/locales/en_US.js
@@ -1,0 +1,3 @@
+export default {
+  "app.title": "3D Editor",
+};