+++ apps/editor-dev/pages/messages/locales/ko_KR.js
@@ -1,0 +1,3 @@
+export default {
+  "app.title": "3D 에디터",
+};