+++ packages/base-editor/src/commands/index.js
@@ -1,0 +1,5 @@
+import keymapKeydown from "./keymap.keydown";
+import keymapKeyup from "./keymap.keyup";
+import toggleTheme from "./toggle.theme";
+
+export default [keymapKeydown, keymapKeyup, toggleTheme];