+++ packages/base-editor/src/commands/toggle.theme.js
@@ -1,0 +1,7 @@
+export default {
+  command: "toggle.theme",
+  execute: function (editor) {
+    var theme = editor.configs.get("editor.theme");
+    editor.configs.set("editor.theme", theme === "light" ? "dark" : "light");
+  },
+};