+++ packages/base-editor/src/configs/index.js
@@ -1,0 +1,3 @@
+import editorTheme from "./editor.theme";
+
+export default [editorTheme];