+++ packages/base-editor/src/utils/detect.d.ts
@@ -1,0 +1,3 @@
+type OSNameType = "mac" | "win" | "linux" | "ios" | "android" | "unknown";
+
+export function os(): OSNameType;