+++ packages/design-system/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @elf-framework/design-system
 
+## 0.0.38
+
+### Patch Changes
+
+- modify tools height, add stripStyle in tabs
+
+- Updated dependencies []:
+  - @elf-framework/design-tokens@0.0.38
+
 ## 0.0.31
 
 ### Patch Changes